refactor: split up large evaluator form file
